+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Yii 2 Crontab extension Change Log
 ---------------------
 
 - Bug #10: Fixed `CronTab::getCurrentLines()` unable to recognize empty crontab at some operation systems (klimov-paul)
-- Enh #7: Added `CronTab::headLines`, allowing setup shell configuration at the crontab beginning (klimov-paul)
+- Enh #7: Added `CronTab::$headLines`, allowing setup shell configuration at the crontab beginning (klimov-paul)
 - Enh #6: In case there is not cron jobs to be saved `CronTab` no longer puts new line separator in result file (klimov-paul)